requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
filelock==3.6.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.14.1
django_environ==0.8.1

djangorestframework~=3.13.1
requests~=2.27.1